requirements-dev.txt
behave==1.2.6
behave-django==1.1.0
django-cors-headers==2.4.0
pre-commit>=1.16.0
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
pytest-forked==0.2
pytest-mock==1.10.0
pytest-testmon==0.9.16
pytest-vcr==1.0.1
pytest-watch==4.2.0
pytest-xdist==1.24.1
tox==3.5.3